+++ Binance.Api/Shared/Converters/BinanceSymbolFilterTypeConverter.cs
@@ -19,5 +19,8 @@ internal class BinanceSymbolFilterTypeConverter : BaseConverter<BinanceSymbolFil
         new KeyValuePair<BinanceSymbolFilterType, string>(BinanceSymbolFilterType.PercentagePriceBySide, "PERCENT_PRICE_BY_SIDE"),
         new KeyValuePair<BinanceSymbolFilterType, string>(BinanceSymbolFilterType.TrailingDelta, "TRAILING_DELTA"),
         new KeyValuePair<BinanceSymbolFilterType, string>(BinanceSymbolFilterType.Notional, "NOTIONAL"),
+        new KeyValuePair<BinanceSymbolFilterType, string>(BinanceSymbolFilterType.IcebergOrders, "MAX_NUM_ICEBERG_ORDERS"),
+        new KeyValuePair<BinanceSymbolFilterType, string>(BinanceSymbolFilterType.MaxNumberOrderAmends, "MAX_NUM_ORDER_AMENDS"),
+        new KeyValuePair<BinanceSymbolFilterType, string>(BinanceSymbolFilterType.MaxNumberOrderLists, "MAX_NUM_ORDER_LISTS"),
     };
 }